% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkData}
\alias{checkData}
\title{Check the entered data against the data dictionary}
\usage{
checkData(dictTable, data, id)
}
\arguments{
\item{dictTable}{a data frame returned by readDataDict}

\item{data}{a data frame returned by readExcelData}

\item{id}{a string indicating the ID variable, to display errors by ID instead of row number}
}
\value{
A list with three data frames: one with all errors, one with errors by row
(or ID if supplied) and one with errors by variable. Also returns a check for duplicate rows.
}
\description{
This function compares the data in the data entry table against the
specifications in the dictTable.
}
\details{
Prior to reading in the data, the dictTable must be imported using
readDataDict and the data must be imported using readExcelData.

The function will check all variables in the dictTable.
If variables are missing from the dictTable an error will occur.
If variables are missing from the data table a warning will be shown.
}
