% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createRecodedVar}
\alias{createRecodedVar}
\title{Create survival variables (survival duration + status)}
\usage{
createRecodedVar(data, dictionary, newVarName, instructions)
}
\arguments{
\item{data}{is data returned by the importExcelData or readExcelData functions}

\item{dictionary}{is the data dictionary returned by importExcelData or readDataDict functions}

\item{newVarName}{the name of the new variable.}

\item{instructions}{are from the data dictionary}
}
\description{
This function will create survival variables from an existing start variable
date of event variable and last date followed variable
}
\details{
The instructions are contained in the Levels column of the data dictionary and should be in the format:

original_varname,newCode1=oldcode1,oldCode2,...,newCode2=oldCode3,..

For Example:
instructions="T_Stage,T0=T0,T1up=T1,T2,T3,T4"

will recode T1-T4 as T1up and retain T0 as is
}
