% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readExcelData}
\alias{readExcelData}
\title{Read Excel Data}
\usage{
readExcelData(
  dataDictionary,
  excelFile,
  dataSheet = "DataEntry",
  range,
  origin
)
}
\arguments{
\item{dataDictionary}{a data frame returned by readDataDict}

\item{excelFile}{path and filename of the data file}

\item{dataSheet}{name of the data entry sheet within the file, defaults to 'DataEntry'}

\item{range}{Optional, Range of Excel sheet to restrict import to (ie. range="A1:F6")}

\item{origin}{Optional, the date origin of Excel dates, defaults to 30 December 1899}
}
\description{
This function reads in an excel data table created by the DataDictionary.xlsm
template file according to the specifications in the dataDictionary
}
\details{
Prior to reading in the data, the dataDictionary file must be imported using
readDataDict.

NOTE: This function will only read in those columns present in the DataDictionary
}
