% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{addFactorVariables}
\alias{addFactorVariables}
\title{Create factor variables from data dictionary}
\usage{
addFactorVariables(dataDictionary, dataTable)
}
\arguments{
\item{dataDictionary}{a data frame returned by readDataDict}

\item{dataTable}{a data frame returned by readExcelData}
}
\value{
a data frame with the updated variables
}
\description{
This function will replace the code and category variables
with factors based on the factor levels provided in the data
dictionary. The original variables are retained with the suffix
'_orig'
}
