% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{checkData}
\alias{checkData}
\title{Check the entered data against the data dictionary}
\usage{
checkData(dataDictionary, dataTable, id)
}
\arguments{
\item{dataDictionary}{a data frame returned by readDataDict}

\item{dataTable}{a data frame returned by readExcelData}

\item{id}{a string indicating the ID variable, to display errors by ID instead of row number}
}
\value{
A list with three data frames: one with all errors, one with errors by row
(or ID if supplied) and one with errors by variable. Also returns a check for duplicate rows.
}
\description{
This function compares the data in the data entry table against the
specifications in the DataDictionary.
}
\details{
Prior to reading in the data, the dataDictionary must be imported using
readDataDict and the dataTable must be imported using readExcelData.

The function will check all variables in the dataDictionary.
If variables are missing from the dataDictionary an error will occur.
If variables are missing from the data table a warning will be shown.
}
