% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createCalculated}
\alias{createCalculated}
\title{Create calculated variables}
\usage{
createCalculated(data, dictTable, timeUnit = "month")
}
\arguments{
\item{data}{is data returned by the importExcelData or readExcelData functions}

\item{dictTable}{is the data dictionary returned by importExcelData or readDataDict functions}

\item{timeUnit}{a string containing the desired unit of time for survival variables}
}
\description{
This function will create survival and recoded variables according to the
rules in the DataDictionary.xlsm file. See the Example sheet for an example.
}
