% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createSurvVar}
\alias{createSurvVar}
\title{Create survival variables (survival duration + status)}
\usage{
createSurvVar(data, newVarName, survVars, timeUnit = "month")
}
\arguments{
\item{data}{A data frame returned by the importExcelData or readExcelData functions}

\item{newVarName}{the name of the new survival variable. The status variable will be suffixed with '_status'}

\item{survVars}{In order the start date, event date and date of last followup}

\item{timeUnit}{Character, the unit of time to calculate survival variables for (ex: 'day' 'week' 'month' 'year')}
}
\description{
This function will create survival variables from an existing start variable
date of event variable and last date followed variable
}
