% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{importExcelData}
\alias{importExcelData}
\title{Import Excel Data from a DataDictionary file}
\usage{
importExcelData(
  excelFile,
  dictionarySheet = "DataDictionary",
  dataSheet = "DataEntry",
  saveWarnings = FALSE,
  setErrorsMissing = TRUE,
  range,
  origin,
  timeUnit = "month"
)
}
\arguments{
\item{excelFile}{path and filename of the data file}

\item{dictionarySheet}{the name of the sheet containing the data dictionary, defaults to 'DataDictionary'}

\item{dataSheet}{name of the data entry sheet within the file, defaults to 'DataEntry'}

\item{saveWarnings}{Boolean, if TRUE and there are any warnings then the function will return a list with the data frame and the import warnings}

\item{setErrorsMissing}{Boolean, if TRUE all values out of range will be set to NA}

\item{range}{Optional, Range of Excel sheet to restrict import to (ie. range="A1:F6")}

\item{origin}{Optional, the date origin of Excel dates, defaults to 30 December 1899}

\item{timeUnit}{character specifying the unit of time that should be used when created survival type variables}
}
\value{
a list containing two data frames: the data dictionary and the data table
}
\description{
This function reads in a data dictionary and data entry table and converts
code and category variables to factors as outlined in the dictionary. This
code is to be used in conjection with the DataDictionary.xlsm template
template file according to the specifications in the dataDictionary
}
\details{
Prior to reading in the data, the dataDictionary file must be imported using
readDataDict.

Warning: If SetErrorsMissing = TRUE then a subsequent call to checkData will not return any errors, because the errors have been set to missing.

NOTE: This function will only read in those columns present in the DataDictionary
}
