% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotVariables}
\alias{plotVariables}
\title{Return a list of univariate ggplots for each non-character variable}
\usage{
plotVariables(data, dictTable, IDvar, vars, showOutliers = TRUE)
}
\arguments{
\item{data}{is a data frame containing the variables to be plotted}

\item{dictTable}{optional, the data dictionary returned by importExcelData or readDataDict functions to provide plot titles}

\item{IDvar}{an optional string indicating the name of an identifying variable to highlight outliers}

\item{vars}{is an optional character vector of the names of variables to plot}

\item{showOutliers}{boolean, Defaults to TRUE. Should outliers be labelled?
Outliers are defined by the 1.5xIQR rule (as with boxplots)}
}
\description{
This function should be run as the final step after the data have been
imported, checked and the factor variables created.
}
