% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotVariables}
\alias{plotVariables}
\title{Return a list of univariate ggplots for each non-character variable}
\usage{
plotVariables(data, IDvar, vars, nAsBar = 6)
}
\arguments{
\item{data}{is a data frame containing the variables to be plotted}

\item{IDvar}{an optional string indicating the name of an identifying variable to highlight outliers}

\item{vars}{is an optional character vector of the names of variables to plot}

\item{nAsBar}{the number of unique values to display as a bar plot for non-factor data, default is 6}
}
\description{
This function should be run as the final step after the data have been
imported, checked and the factor variables created.
}
